[assembly] Generate devmgr config in the image assembler
This is required in order to transition fully from the GN assembly code
to the image assembler. The devmgr_config.txt file is now generated
inside of the image assembler and added to the list of BootFS files. For
bringup (bootfs-only) builds, the devmgr_config.txt file is empty.

Bug: 79674
Change-Id: I036c2cc684599c078f4c30555f5aa2075d5c16a5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/549563
Reviewed-by: Aaron Wood <aaronwood@google.com>
Commit-Queue: Aidan Wolter <awolter@google.com>