[componentsv2] (AbsoluteMoniker|Relative)Moniker support partial monikers
We would like to support binding with partial monikers that don't
include instance IDs as this is a runtime detail that isn't easily known
to developers. In order to do so we need support for
PartialAbsoluteMoniker and PartialRelativeMoniker. This CL introduces
both. We do this by making AbsoluteMonikerBase and RelativeMonikerBase
traits with associated types and implement concrete versions in
AbsoluteMoniker, ParitalAbsoluteMoniker, RelativeMoniker and
PartialRelativeMoniker.

Change-Id: I2f1e6ebd87d26236c245ba807522e1d557e3b1f1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/553061
Commit-Queue: Fady Samuel <fsamuel@google.com>
Reviewed-by: Gary Bressler <geb@google.com>