[packet] Improve ergonomics of FragmentedBytes
FragmentedBytes{Mut} used to have a single lifetime parameter for both
levels of references. However &mut T is invariant over T, this makes it
inconvenient to work with when we want the outer borrow to have a
shorter lifetime, transmute can be used as a workaround but after all,
it is unsafe. One more lifetime parameter is given to the structs, so
that they can be unified independently.

While I am here, anonymous lifetimes are removed in the where clauses
to not make readers confuse it with lifetime elision. Lifetime elision
in where clauses is not supported and its meaning is not even decided
(https://github.com/rust-lang/rust/issues/45667). To avoid surprises and
to be more specific, we ought to change those to use the explicit
desugared HRTB form.

Test: No behavior change
Change-Id: I8b66d95822cc261fdbdd0be2d686da4a1194ca9b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/552743
Commit-Queue: Zeling Feng <zeling@google.com>
Reviewed-by: Joshua Liebow-Feeser <joshlf@google.com>