[starnix][pipes] Implement blocking partial writes
When writing to a pipe in blocking mode, if there is a partial write,
then the operation will continue to block waiting for more space to
become available in the pipe. If the reader shuts down, then the write
completes as a short write an the caller receives a SIGPIPE signal.
This CL updates our behavior to match the description above instead of
simply returning a short write immediately.

After this change, we pass all but a few tests in pipe_test. This CL
switches over to a denylist of tests we do not pass instead of the
previous allowlist of tests we do pass.

The remaining tests either require procfs, statfs, or O_LARGEFILE.

We now pass 114 tests in this test suite.

Test: PipeTest.BlockPartialWriteClosed, ReadFromClosedFd, Streaming
Change-Id: I5e2cf64e345a0f2c5ac3624741087cd29e7db460
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/555753
Fuchsia-Auto-Submit: Adam Barth <abarth@google.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Commit-Queue: Adam Barth <abarth@google.com>