[kernel][virtual_alloc] Add virtual allocator
The virtual_alloc library implements an allocator that can be given a
range of virtual memory and then provides a page granuale allocator
inside that. It has no dependencies on the heap and only uses the pmm
and MMU (which also have no heap dependencies).

This commit just adds the library but leaves it unused.

Bug: 30973
Change-Id: Ia020cc11fea1a953c6ce4f4f243c8a2262b42c04
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/514687
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>