[cmpctmalloc] Increase maximum heap growth to 2MiB
This changes the maximum heap growth amount from 1MiB to 2MiB.

The intention of growing by 2MiB, instead of 1MiB, is so that when
using a virtual heap, that requires mapping in new pages, large
pages can be used, saving on MMU memory usage.

In this worst case this could increase memory usage of the heap by
1MiB. However, this requires the heap to have already grown such
that 2MiB is 12% (the heap grow ratio) of the heap at which point
an extra 1MiB is a relatively small portion of memory.

Bug: 30973
Change-Id: I1f3a2c5172f46021705bc102d35d79591b7262ff
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/528795
Fuchsia-Auto-Submit: Adrian Danis <adanis@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>