[build] Remove all_outputs_fresh from action() wrappers.
This extra argument to action() and action_foreach() is no longer
relevant. Its default value is 'false', and all use cases also
set it to 'false'.

This CL removes the associated code paths, as well as use cases
in fuchsia.git. There are however still a few third_party/...
BUILD.gn files that set the variable when calling action() or
compiled_action(), so ensure these still build properly. However,
use an assert to disable setting the variable to 'true'.

The //third_party/... files that will need to be updated are:

  //third_party/crashpad/BUILD.gn
  //third_party/cobalt/build/config/BUILDCONFIG.gni
  //third_party/cobalt/metrics_registry.gni

Bug: None
Change-Id: Ic01141c42b5177b296b4f82615f6c1c93c1a6e79
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/556002
Commit-Queue: David Turner <digit@google.com>
Reviewed-by: Jay Zhuang <jayzhuang@google.com>
Reviewed-by: David Fang <fangism@google.com>