[iwlwifi] Don't require a newline in diagnostics
Remove runtime check that triggers an assert when a diagnostic
format string doesn't contain a newline. In fuchsia, zxlogf()
doesn't require a newline (and will in fact generate a blank
line if we use one), so we strip it from diagnostics.

This CL simply removes the code that forced the runtime assert()
when a newline wasn't present.

Bug: 77650
Test: Manually verified that a diagnostic without a newline doesn't
      cause a crash. Also ran a quick smoke test to verify that
      things are generally working as before.
Change-Id: I3522fa3ae7896e5124485b36558ae083192cb326
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/556183
Reviewed-by: Louis Yung-chieh Lo <yjlou@google.com>
Commit-Queue: Joshua Conner <joshconner@google.com>