[mold] Replaced bit set with a range set.
The old implementation of the set of composition IDs was backed by
a bit set since it was optimized for memory only. This led to
somewhat poor performance (O(N)) when using quite a lot of layers.
The bit set is being replaced in this CL with a range set that is
pretty memory efficient, but finds new layer IDs in constant time.

Testing: fx test mold-tests
Change-Id: I67b167def9a51a5cdf904d2857e8fdccdc222177
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/556381
Commit-Queue: David Reveman <reveman@google.com>
Reviewed-by: David Reveman <reveman@google.com>