[docs][diagnostics] Make bitfield diagrams more clear.
It is more useful to see the start index of each space than the index of
the following space. This CL changes the numbering to instead show start
index, and we update the description document to be more clear about how
to interpret the new format.

Change-Id: If9383c2caf2a8f945cb53f653f4a933785c35fb1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/556405
Fuchsia-Auto-Submit: Christopher Johnson <crjohns@google.com>
Reviewed-by: Miguel Flores <miguelfrde@google.com>
Commit-Queue: Christopher Johnson <crjohns@google.com>