[mold] Replaced DashMap.
The DashMap used in mold was more of a commodity and not strictly
required given the requirements. However, since in high-resolution
environments it was adding even 1ms+ of latency to drawing, it is
being replaced with a Vec that can be used in parallel by taking
advantage of rayon's functionality.

Testing: fx test surpass-tests && fx test mold-tests
Change-Id: I43032404cfe32e9780ea9f149b73f6d3899b2c49
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/556341
Commit-Queue: David Reveman <reveman@google.com>
Reviewed-by: David Reveman <reveman@google.com>