[host][blobfs] Convert depfile to buffered FILE*
The host blobfs tool reads files, hashes them, and compresses them in a
small collection of threads. Each thread takes a blob to process from a
global queue, processes it, and adds back the hashed and compressed
object to a completed queue.

The tool may optionally output information about each blob to a
'depfile', a standard file with each source blob listed. Each thread
writes to the depfile as it processes blobs, which required some
cross-thread synchronization.

Convert the depfile from a mutex + raw FD to a buffered FILE*.
The FILE*'s fwrite() path only writes to an in-memory buffer, which
may be flushed later; this removes a syscall-under-a-mutex that
every thread would have to pay for and batches writes to the
depfile fd.

As part of fxr/404773, this change and the other parts of that fxr/
reduced runtime of 'blobfs create' by ~2% when packing core.x64
--with-base //bundles/buildbot:core on a 16-vCPU VM on GCE.

Bug: 55297 It would be nice if the Fuchsia build was faster
Change-Id: Ic60ca9d8025553fecd594be05e2b9409c83aefd9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/405102
Testability-Review: Shai Barack <shayba@google.com>
Testability-Review: James Sullivan <jfsulliv@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>
Reviewed-by: Shai Barack <shayba@google.com>
Commit-Queue: Shai Barack <shayba@google.com>