[amlogic-display] Change vsync thread RDMA synchronization
In the absence of RDMA irqs, it is possible for the vsync thread to
not wake up while waiting for the condition variable in
WaitForRdmaIdle(). It now uses a timed wait to alleviate this.

The timeout means that the RDMA blocking condition now executes more
frequently. This logic has also been changed to ensure that that wait
condition only depends on the state of the auto-RDMA done interrupt
status and the loop itself only entered if an RDMA request is known to
yet be serviced and the RDMA interrupt done bits haven't been cleared
yet.

Also ran fx format-code.

Test: 1. Inject an artificial 20 second sleep in RDMA thread. Confirm
         that the display only updates every 20 seconds.
      2. Observe no register dumps in the logs without the TimedWait
         change.
      3. Observe register dumps with the TimedWait change.
      4. Ensure that the display continues to refresh.
Change-Id: Idff4063f9f2cf6aa7e72286c9927d7150d981480
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/556171
Reviewed-by: Randall Bosetti <rlb@google.com>
Reviewed-by: John Bauman <jbauman@google.com>
Commit-Queue: Arman Uguray <armansito@google.com>
(cherry picked from commit 10ddc39c010a0538e6d035c66214a6de8805646e)