[starnix] Pass ExitTest.CloseFds
We now represent zombie tasks with a minimal struct that contains the
information we need rather than a complete Task object. This approach
lets us drop the Arc<FdTable> when the Task exits as opposed to when the
zombie is collected, which is required to pass ExitTest.CloseFds.

After this CL, we pass all the tests in exit_test.cm.

Test: ExitTest.CloseFds
Fixed: 79266
Change-Id: Id03d66707ac058f10adc1dcc3b1e7f326cdc16a8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/556113
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Adam Barth <abarth@google.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>