[display] Rename primary/vc nodes as client
Renamed the `primary` and `vc` nodes to the more generic term
`client` and added a "primary" boolean property to convey that
semantic meaning. The more generic term allows a selector such as
"client-*" to filter for all clients, at the expense of the ability
to specifically filter for primary or virtcon.

Change-Id: I5134bbbf9a137eff2dbf723dcce31344900cb45f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/555488
Reviewed-by: John Bauman <jbauman@google.com>
Reviewed-by: Randall Bosetti <rlb@google.com>
Commit-Queue: Arman Uguray <armansito@google.com>