[size_checker] Check the Update Package size when absent from blobfs
The Update Package is being removed from blobfs, but we need to continue
verifying that the size is within budget. We are adding a separate
section to the size_checker JSON that allows us to verify the size of
blobs even if they are not assembled into BlobFS.

This verification is handled through use of the blobfs-compression tool,
which does not de-duplicate blobs, and sometimes over-estimates the
size. This is alright for the Update Package, because the amount saved
by de-duplication is negligible (~9kib).

Bug: 58645
Test: fx test size_checker_tests
Change-Id: I09f71a341d065e5b5e28aee2fd65e731d4f6c760
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/554320
Commit-Queue: Aidan Wolter <awolter@google.com>
Reviewed-by: Fady Samuel <fsamuel@google.com>