[bt][mock-piconet-server] Fix bug in generating ServiceFoundResponse
The ServiceFoundResponse for protocols that specified RFCOMM should
include the RFCOMM channel number so that peers in the piconet can
correctly connect to the service. This was incorrectly removed in
fxrev.dev/553171 in which all RFCOMM-related parsing was removed.

Update the local ServiceRecord type to store the primary protocol
directly. This simplifies building the ServiceFound response and
prevents any sort of information loss in the conversion process.

Update test helpers to validate `peer_id` in ServiceFound responses.

Bug: 79742
Test: fx test mock-piconet-server-tests
Change-Id: Id47cde9267edbe977c1ec83b02d49e01199067ca
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/555943
Commit-Queue: Ani Ramakrishnan <aniramakri@google.com>
Fuchsia-Auto-Submit: Ani Ramakrishnan <aniramakri@google.com>
Reviewed-by: Lucas Jenkins <lucasjenkins@google.com>