[carnelian] Use BTreeMap for compositions.
This makes it possible to change the number of layers in a
composition while maintaining a stable order. Critical for
good text grid performance.

This reverses the order of some scene building code to
become top to bottom to match the btreemap.

The textgrid example has been updated to take advantage
of this and supports partial updates as a result.

Test: carnelian-tests
Change-Id: Iecd4f3cd75ae7a508aa49418650aaf1e778a538e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/556321
Commit-Queue: David Reveman <reveman@google.com>
Reviewed-by: Dragos Tiselice <dtiselice@google.com>
Reviewed-by: Rob Tsuk <robtsuk@google.com>