[build] Integrate verify_routes into the Fuchsia build.
This change integrates the ffx scrutiny verify routes command into the
build so that the build will fail if a use statement from a protocol or
directory is introduced without a matching expose-offer chain in the
build. This is to prevent invalid routes from being added to the image.

This does not include components that are not linked up at some point to
root.cml and only verifies v2 components.

Change-Id: I6f5062891189692cf1cdbadc82915f94cb4f8083
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/545001
Commit-Queue: Benjamin Wright <benwright@google.com>
Reviewed-by: Gabe Schine <thatguy@google.com>
Reviewed-by: Shai Barack <shayba@google.com>