[cpp-realm-builder] Add nullptr check, auto name.
This CL adds a nullptr check to all functions/methods,
internal and public, that receive a pointer type.
Also, it adds a static method overload for ScopedInstance::New()
that uses an auto-generated child name for a dynamic child.
Finally, the RealmBuilder library uses such child name for the
root of the created Realm.

Change-Id: I7da296fd2a4327f168058b2c7be40e8c3bb8fb68
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/556107
Commit-Queue: Yaneury Fermin <yaneury@google.com>
Reviewed-by: Gary Bressler <geb@google.com>
Reviewed-by: Miguel Flores <miguelfrde@google.com>
Reviewed-by: Gabe Schine <thatguy@google.com>