[power] Selective CPUs in CpuControlHandler
Adds support in CpuControlHandler to use a subset of all the CPUs in the
system for determining load. This lets the CpuControlHandler node inform
its power calculations based on just the CPUs that are under its
purview.

There should be no functional change in behavior caused by this change.

Bug: fxbug.dev/50974
Test: fx test power-manager-tests
Change-Id: I14e4e96a000862367d25fe0a344bef031793d419
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/556565
Commit-Queue: Patrick Shickel <pshickel@google.com>
Reviewed-by: Onath Claridge <claridge@google.com>