[blobfs] Support for multiple pager threads.
This commit does not enable multiple pager threads nor does itself
guarantee that we're ready to use multiple threads, but it does provide
a path for turning it on. This will only actually enable multiple
threads for the "new" pager. On the old pager it will allocate extra
resources that will never actually be used.

Adds a mount option for number of paging threads, and uses that to set
the pager threadpool size and allocate the number of required resource
sets for multiple threads to work independently. These resource sets are
then assigned to each pager thread on their first call into the pager by
storing the thread's worker id into a thread local.

Bug: 67659

Change-Id: I477fb28225f87dce263b1824eb718a1806ff58a4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/552537
Commit-Queue: Martin Lindsay <mlindsay@google.com>
Reviewed-by: Brett Wilson <brettw@google.com>