[assembly] Complete removal of product_include_updates_in_blobfs
product_include_updates_in_blobfs deprecation step (4).

We are in the process of removing product_include_updates_in_blobfs.
This flag is used to put the update package in blobfs at assembly time,
which ensures that the device can take an OTA by reserving that space.

Now that we have the max_blob_contents_size flag, we can be confident
that we will not go over budget.

The deprecation is as follows:

(client repos)
1. Explicitly set the flag on all smart products to false &&
   Reduce budget for blobfs to ensure OTAs are possible
   (this removes update package from blobfs safely)

fuchsia.git
2. Change the default state of the flag to `false`
   (allows us to remove usage of the flag in clients)

(client repos)
3. Remove usage of the flag

fuchsia.git
4. Remove the flag and usage in fuchsia.git

Bug: 58645
Change-Id: Ic13582b5746bd4bf5defc72a9a3b99061672e473
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/549964
Commit-Queue: Aidan Wolter <awolter@google.com>
Reviewed-by: David Turner <digit@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>