[build][go] Ensure the right version of 'ar' is used for C archives
It was discovered by running our build in a container with a minimal
filesystem image that the "go build" command will invoke the host
'ar' tool when building a c-archive library. This failed because
there was no 'ar' in the PATH in the container environment.

This change ensures that the correct toolchain-specific 'ar' is
used instead in this case.

Bug: None

Change-Id: Ia8673154a98118eac6458399541c3901a1b1823c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/552183
Reviewed-by: David Fang <fangism@google.com>
Reviewed-by: Jay Zhuang <jayzhuang@google.com>
Commit-Queue: David Turner <digit@google.com>