[netstack] Log when interface watcher queue full
Log a warning when a `fuchsia.net.interfaces/Watcher` is closed
because the event queue is full. This may be because the client
is no longer interested in the events and have stopped calling
`Watcher.Watch`, in which case it should have closed the channel;
the server is generating spurious events or unexpectedly introducing
latency while handling the requests; or some environmental factor
such as unexpected latency anywhere in the FIDL method call chain.

Also lower the maximum queue length to 128 from 512 (which was
excessively conservative).

Change-Id: Iba8772ea3c8d443c6005b57eedfb93935dc5b3ab
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/557050
Commit-Queue: Tony Gong <gongt@google.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>