[kernel][arch] try to use const accessor on more cases where Thread is being passed as args
Started as a small clean up to add a const accessor for the arch() field
of Thread and expanded out to a lot of call sites that could be passing
a const Thread * into a function but arent.

No functional change. Just some more constness.

Change-Id: I2c20d7865b18ae5b1602459d548fec1bbe350246
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/556741
Reviewed-by: Adrian Danis <adanis@google.com>
Commit-Queue: Travis Geiselbrecht <travisg@google.com>