[kernel][crashlog] Add "persistent" dlog support
Add the ability to share some of our Very Special Crashlog memory in
order to produce a persistent dlog.

The persistent dlog is a ring buffer into which we always store the
most recent messages which get written to the dlog. Data written are
flushed from cache to RAM each time, allowing them to be recovered and
added to a crashlog in the case of a spontaneous reboot. Writes will
never block waiting for space, or waiting for a UART. Older data are
simply overwritten.

Tests: new kernel unit tests
Bug: 70724
Change-Id: Ica479d5e16390a38e9aa8afb631aaa5499336a6b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/506930
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Reviewed-by: Corey Tabaka <eieio@google.com>
Reviewed-by: Carlos Pizano <cpu@google.com>
Commit-Queue: John Grossman <johngro@google.com>