[kernel][arm64] Ensure EL2 page tables are disabled during mexec.
We disable caches and the MMU for EL1, but fail to do so for EL2.

If a guest happened to be running when we call mexec, we will attempt to
jump into the new kernel with the MMU and caches enabled, violating the
assumptions of the new kernel (and possibly writing over the page tables
while we shuffle data around for the new kernel).

Update mexec to also disable the MMU and caches for EL2 if we're using
it.

Change-Id: Ic5ff166fc9b3fff16c3fd3fcde4c48d73bb08774
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/550927
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Commit-Queue: David Greenaway <dgreenaway@google.com>