[fidl][rust] Bump Rust parameter limit from 11 to 12
This CL makes Rust support FIDL libraries with methods that have up to
12 parameters. Previously it supported only up to 11.

This issue arises from the fact that Rust bindings use tuples
request/responses rather than generated structs, and it is not possible
to implement traits for an arbitrarily sized tuple.

This also improves the situation by reporting an error from fidlgen_rust
when the limit is exceeded, rather than confusing the user with a
cryptic rustc error about Encodable not being implemented on a type.

Test: fx test //src/lib/fidl/rust
Test: added golden methods with 11 and 12 parameters, latter denylisted for Rust
Bug: 76655
Change-Id: I9f0150f3a51803775815ffe371f7eb32a06e7886
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/556666
Commit-Queue: Mitchell Kember <mkember@google.com>
Reviewed-by: Benjamin Prosnitz <bprosnitz@google.com>