[memalloc] Introduce Pool::UpdateFreeRamSubranges()
This method converts all kFreeRam subranges of any given range to a
given, extended type.

The utility of this operation lies in situations where there is a
special range that we ultimately want reserved for "something" later on,
but it is immaterial what occupies it in the meantime, so long as
nothing is further allocated from there. This will be required in the
case of loading a fixed-address kernel image, where we would want to
prevent page tables - which must persist across the boot - from being
allocated out of the that load range.

Bug: 77359
Change-Id: I8e2e2a153b77ba09f69557891e77512e667c8263
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/545948
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>