[component_manager] Implement reboot-on-terminate
This change implements the 'on_terminate: reboot' feature for parity
with v1 critical_components, as described in the linked bug.

This introduces some changes to the model to implement the feature. The
top instance (component_manager's instance) now has a state object to
hold a reference to its child (i.e., the root) and the reboot task, if
in progress. The child pointer allows the top instance to access the
statecontrol.Admin protocol exposed from the root.

This change does not enable the feature, so clients may not use it yet.

Bug: 80365
Change-Id: I530ed79443c8296a5f8c741d0e5ddc91cd00c6d1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/553607
API-Review: Gary Bressler <geb@google.com>
Commit-Queue: Gary Bressler <geb@google.com>
Reviewed-by: Justin Mattson <jmatt@google.com>