[flatland] add traces for vsync->OnFramePresented
While investigating fxbug.dev/81117, I found that the delay was simply
caused by unfinished behavior in the Flatland display compositor.
Completing the implementation should lead to the right behavior. In the
meantime, this CL changes the wakeup from 16.7ms in the future to 14ms,
which is also guaranteed to be correct.

In any case, it's worthwhile to add these traces for future reference.

Bug: 81117
Test: tested with FlatlandViewProvider
Change-Id: I2f638052e7535ea34f58c0d289abc545eefc1cc1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/557981
Commit-Queue: Felipe Archondo <felipearchondo@google.com>
Fuchsia-Auto-Submit: Felipe Archondo <felipearchondo@google.com>
Reviewed-by: Josh Gargus <jjosh@google.com>