[lazy-init] Remove address-of operator.
Remove the address-of operator from LazyInit<>, replacing it with a
specific method (GetAddressUnchecked) instead.  The primary purpose of
overloading it in the first place was to allow code to take the
address of the underlying T at compile time to do things like
initialize a global array of pointers.  The runtime checks present in
the Get() method would prevent this in some cases.

Instead, users may now use GetAddressUnchecked to explicitly fetch the
address of the underlying storage without hitting the runtime checks.

Also, fix a number of places where the code was inappropriately using
the & operator to fetch the address of the underlying storage, and
accidentally bypassing any checks in the process.

Test: existing unit tests.
Change-Id: I61a302155336ba369d7a7c60de045f9c27cdebcc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/557801
Reviewed-by: Corey Tabaka <eieio@google.com>
Commit-Queue: John Grossman <johngro@google.com>