[kernel][io] Refactor Linebuffer helper.
Move the Linebuffer helper class out of lib/io and into the Thread
class as an inner struct.  This allows the storage details of the
linebuffer to be a property of threads while leaving the printing
behavior as a private detail of the io library.

Change-Id: I03dd38933260333847f2504a764e56020b6ef0b3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/555744
Commit-Queue: John Grossman <johngro@google.com>
Reviewed-by: Carlos Pizano <cpu@google.com>