[docs][fidl] Fix links in fidl documentation
Fix how links to source code and live documentation
are written in comments in fidl files so that they
render and navigate correctly both locally and in
fuchsia.dev.

All the links fixed in this CL are currently broken
when browsing the corresponding
https://fuchsia.dev/reference/fidl pages.

There is currently no standard for links to docs in
fidl and code comments regarding being absolute or
relative. Currently, they need to be host and path
absolute in order to render correctly in fuchsia.dev.

This CL is not an attempt to define such a standard,
and a definitive decision should result from an
ongoing discussion in Fuchsia Devrel and Tech
Writers.

Change-Id: I721f7e6469cba335392c002c5f7870e8cd45829c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/557724
Reviewed-by: Andres Oportus <andresoportus@google.com>
Reviewed-by: Marie Janssen 💖 <jamuraa@google.com>
Reviewed-by: Gabe Schine <thatguy@google.com>
API-Review: Marie Janssen 💖 <jamuraa@google.com>
Commit-Queue: Renato Mangini Dias <mangini@google.com>