[bt][sm] Don't reset timer when rejecting pairing request
As initiator, we reject any Pairing Requests sent to us by the peer
from the top-level SecurityManager without ever enter a pairing phase.
This caused issues because the SM Channel method we used restarted the
SM timer, so the timer could expire without being in a pairing phase,
violating the invariant that a timer is never active while not pairing.
Fix this by using the variant of the method that does not restart the
timer.

While here, update the Responder to reject Security Requests in a
manner symmetric to how the Initiator rejects Pairing Requests.

Fixed: 80826
Test: Added SMP_InitiatorPairingTest.RejectPairingRequest and
      SMP_ResponderPairingTest.RejectSecurityRequest to `$ fx test \
      bt-host-sm-tests`
Change-Id: I24f1305ac49c4efb28adeeb8406b0b68920c4bee
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/556285
Fuchsia-Auto-Submit: Lucas Jenkins <lucasjenkins@google.com>
Reviewed-by: Ben Lawson <benlawson@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>