[diagnostics] Create diagnostics-kcounter that starts on boot
This is a re-implementation of src/developer/kcounter_inspect in Rust
using the unified diagnostics launcher. It publishes the kcounter
Inspect VMO from kernel as lazy values at the root of the component's
Inspect tree, and updates the VMO whenever it is accessed.

Test:
* A unit test verifies that we call {get,update}_inspect_vmo correctly
  and lazily.
* An Inspect integration test is added to verify the existence of
  handles info in the Inspect tree produced by this new component,
  similar to the integration test for kcounter_inspect.

Start-on-boot is achieved via an empty control protocol that's consumed
by archivist, similar to diagnostics-persistence.

Bug: 79123
Change-Id: I970352aaccf1cea9c4ae61c469bd163d26b2a182
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/555362
Commit-Queue: Robert Ma <robertma@google.com>
Reviewed-by: Miguel Flores <miguelfrde@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>
Reviewed-by: Shai Barack <shayba@google.com>