[audio_core] Increase memory-pin frequency
This CL changes AudioCore kTimeBetweenPins from 15s to 5s.

When the system is under memory pressure (Warning and Critical),
the kernel evicts pager backed memory that has not been accessed
in the last 30s. But when system memory falls below the OOM
threshold, the kernel evicts more aggressively, evicting memory
that has not been accessed in the last 10s.

To ensure that audio never misses a deadline because of audio-
related paging, we periodically pin audio_core executable memory
(slightly less than 4MB in the relevant scenarios). This duration
was previously 15s and is now 5s.

Fixed: 81266

Change-Id: Icae71f0fa934ce0c82ca84fd445bbf6354f56285
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/558801
Reviewed-by: Austin Foxley <afoxley@google.com>
Commit-Queue: Martin Puryear <mpuryear@google.com>