[fxfs] Fix race between unlink and open
We need to take an open count whilst we hold locks during lookup since
otherwise it's possible for a file to be unlinked and then purged after
lookup but before the open count has been increased.

The other issue was with purging from the node cache. We need to check
that the node we are removing from the cache is the one we expect.

Fixed: 80975
Change-Id: I7ac89580c872353a7891cf794d1ce16d1b318883
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/558007
Commit-Queue: Chris Suter <csuter@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>