[fvdl] Replace use of "curl" with hyper's "get"
In fxr/537543, a comment was made to replace a std::process::Command invoking
the "curl" executable with a call to hyper's "get" command. This change was made,
but a bad commit/upload overwrote it, and so the "curl" command was pushed to
HEAD.

This CL redoes the change from "curl" to "get".

Change-Id: I72e1830f829d419e2499b2073843ed999997bed0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/559022
Commit-Queue: Daniel Dougherty <doughertyda@google.com>
Reviewed-by: Yuan Zhi <yuanzhi@google.com>