[rbe] un-hardcode env var input files
Let GN files be the source of dependency-truth, and make specifying
additional input dependencies (remote execution uploads) maintainable
moving forward.
Without this fix, developers have to maintain redundant information in
two places, and is prone to breakage, and unpleasant surprises in CQ.

Method:
Pass an extra dummy rustc flag that lists inputs that need to be
uploaded for remote building.  This flag is only meaningful to the
rust-remote-wrapper.sh script and is stripped away before local and
remote execution.
The --remote-inputs flag is comma-separated, and makes its way to
the list of (rewrapper) --inputs to upload.

Bug: 79963
Bug: 81311

Change-Id: I8c718d5a8b9ed9be2e1303f0c7eb42efa77465b6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/554318
Commit-Queue: David Fang <fangism@google.com>
Reviewed-by: David Turner <digit@google.com>