[debugger] Correct PC for hardcoded breakpoints
Previously, debug_agent will set PC to the next instruction after the
breakpoint instruction for hardcoded breakpoints, e.g. in abort(). This
creates confusion for users and the unwinder. Change this behavior so
that the adjustment of PC is done on resuming from the exception.

Change-Id: I005c19958e0fb4433d9f38585d68b56c83ffb71a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/558401
Commit-Queue: Dangyi Liu <dangyi@google.com>
Reviewed-by: Brett Wilson <brettw@google.com>