[i2c][amlogic] Don't log errors for NACKs
According to the I2C spec there are valid reasons for devices to NACK
bytes from the host. For example, the SHTV3 sensor on Nelson NACKs
requests to read temperature data when a measurement is in progress.
Remove the error message in the Amlogic I2C driver to eliminate log spam
in these cases, and instead rely on client drivers to log when
unexpected errors occur.

Fixed: 81383
Test: No more I2C error messages on Nelson
Change-Id: Ia670918f0b5f25af549d194123533ae9c1716d99
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/560367
Reviewed-by: Eric Holland <hollande@google.com>
Reviewed-by: Andres Oportus <andresoportus@google.com>
Commit-Queue: Braden Kell <bradenkell@google.com>