[blobfs] Change the default blobfs format.
The default blobfs format in the code when creating a new format
structure was the deprecated padded format. This changes the default.

Renames the enum to "kDeprecated..." so it's more clear at any call
sites that this should not be used.

Convert the tests to explicitly specify a padded merkle tree where
before they specified a compact one. The rest of the tests should now be
using the compact format.

There should be no production behavior change as all build commands that
produce images explicitly specific the desired format.

Change-Id: Id6c42a9b7ecade6989bf377b80575ac080b980a2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/559185
Reviewed-by: Chris Drouillard <cdrllrd@google.com>
Reviewed-by: James Tucker <raggi@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>