[fidl-lint] Remove repeater checks
Repeater checks attempt to enforce the aspect of the FIDL style guide
that says to avoid repeating a component from the parent context's
name (unless it serves to disambiguate).

This check has never been enabled due to false positives - given the
amount of time this feature has existed but not been touched,
we opt to remove this feature rather than continue maintaining it
(for example, updating it to support anonymous layouts). This is
done by simplifying the context stack from a deque<Context> into
a stack<string> representing the layout kinds (the deque -> stack
is an independent change to improve readability).

Test: fx ninja -C out/default host_x64/fidl-compiler &&
./out/default/host_x64/fidl-compiler
--gtest_filter='LintFindingsTests.*'

Change-Id: I26d5c869e5484b0928e7306bcba58b17febadeb4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/560443
Reviewed-by: Ian McKellar <ianloic@google.com>
Commit-Queue: Felix Zhu <fcz@google.com>