[fxfs] Fix race in graveyard reaper
When the reaper task finished it would remove its own future, but
that would mean that if wait_for_reap was called just at that moment,
it wouldn't wait, and would just continue.  The problem is that at
that point the Arc reference hadn't been dropped, but test code
depends upon this. The solution is to not remove the future when the
task finishes.

Also, the paused state wasn't doing anything so I got rid of that and we
can just use an Option instead.

Fixed: 80777
Change-Id: Icf52e90dde734fbf24dff62ee0e16ca7198637ea
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/559902
Commit-Queue: James Sullivan <jfsulliv@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>