[build][tarmaker] Sort entries to make output more deterministic
The 'tarmaker' implementation uses a map[string]string container
that has randomized key order during iteration. This resulted in
non-deterministic outputs every time the tool was run with the
same inputs.

This fixes the issue by sorting the entries by path in the archive
before generating it.

Note that the output is not fully deterministic yet, because the
file timestamps and user/group values are not normalized at the
moment.

Fixed: 81322
Change-Id: Ib57d2efcc589179865b9f06706047fc983663f0c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/559204
Commit-Queue: David Turner <digit@google.com>
Reviewed-by: David Fang <fangism@google.com>
Reviewed-by: Jay Zhuang <jayzhuang@google.com>