[blobfs] Update blob command-line flags
Changes the blobfs tools from the flag
  --blob_layout_format=[padded|compact]
to
  --deprecated_padded_format
when the old format is needed, and nothing when it's not. This is to
help document that this format will not be supported forever, and to
encourage the compact format by making it the default.

Both flags are currently supported in zircon/tools/blobfs because there
are some out-of-tree scripts that use the existing flag.

There should be no behavior change as all builds should get the same
flag they had before, only the way it's specified has changed.

Bug: 79775
Change-Id: I14d0364987603ef9c945b5a4f8c71d7d3d61a687
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/559191
Reviewed-by: Aidan Wolter <awolter@google.com>
Reviewed-by: Chris Drouillard <cdrllrd@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>