[assembly] Allow clients to specify the output ZBI name
The board config for the Image Assembler now uses board.zbi.name to
allow clients to specify the output ZBI name. "fuchsia" for
"fuchsia.zbi". A "signed" suffix is appended if a ZBI signing tool is
specified in the board config.

The Image Assembler generates two ZBIs:
1) <board.zbi.name>.zbi
2) <board.zbi.name>.zbi.signed (if signing tool is specified)

The name of the ZBI as mentioned in the Update Package is:
1) zbi
 OR
2) zbi.signed (if signing tool is specified)

Bug: 79886
Change-Id: I42f31fbab9535362b7a54e08e2ac42a55ee19caf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/559886
Commit-Queue: Aidan Wolter <awolter@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>