[fuchsia-async] Fix deadlock in emulated channel
The emulated zircon handle implementation shards the process global
handle table to reduce lock contention.  It also needs to be careful to
not re-entrantly lock the same shard, or that operation will deadlock.
A test flake revealed a rare edge case which may end up in this
situation.

When sending a channel through a closed channel, the write_etc call will
detect that the outer channel is closed and close all handles that were
given to the write_etc call, however, it was doing that with the outer
channel's shard locked.  Closing a handle also locks the shard for that
handle, so if both channels were in the same shard, that write would
deadlock.

This change modifies the error path to defer closing the handles until
after it has released the shard's lock, avoiding the deadlock.

Looking through other APIs here, this seems to be the only situation
where this kind of deadlock could occur.

Fixed: 81561
Change-Id: Ief8892867ddce2803640b67e9768b4db2827e562
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/561130
Fuchsia-Auto-Submit: Kevin Wells <kevinwells@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>